move stories into component folders
