+++ frontend/src/app/commonComponents/CardBackground/CardBackground.tsx
